+++ src/def/IAccountJailInfo.d.ts
@@ -1,0 +1,4 @@
+interface IAccountJailInfo {
+    // Contains the date with lock action expire
+    endLock : number;
+}