+++ src/def/IUserJailInfo.d.ts
@@ -1,0 +1,4 @@
+interface IUserJailInfo {
+    // Contains the date with ban action expire
+    endBan : number;
+}