+++ test/functional/protocols/local.js
@@ -11,14 +11,15 @@ describe('Functional :: Protocols :: Local', function () {
     it('should have *local* protocol selected', function() {
         assert(trapjs.getProtocolName() == 'local', 'error must get local but  : '+trapjs.getProtocolName());
     });
+
     it('should simple user attempt', function(done) {
         trapjs.loginAttempt('Seby45', '112.0.0.0', function(err) {
             assert(err == void 0, 'error must be undefined : '+err);
             done();
         });
     });
-    it('should ban user', function(done) {
 
+    it('should ban user', function(done) {
         // Loop login for ban
         for (var i = 0; i < 11; i++) {
             trapjs.addAttempt('Seby45', '112.0.0.0');
@@ -55,6 +56,10 @@ describe('Functional :: Protocols :: Local', function () {
             assert(err.code == 'E_USER_BAN');
             assert(typeof err.user.banTime == 'number');
             assert(err.user.banTime > 7150 && err.user.banTime < 7201);
+
+            // Unban user
+            trapjs.unbanUser('112.0.0.0');
+
             done();
         });
     });
@@ -69,6 +74,28 @@ describe('Functional :: Protocols :: Local', function () {
             assert(err.code == 'E_USER_BAN');
             assert(typeof err.user.banTime == 'number');
             assert(err.user.banTime > 86300 && err.user.banTime < 86401);
+
+            // Unban user
+            trapjs.unbanUser('112.0.0.0');
+
+            done();
+        });
+    });
+
+    it('should lock account', function(done) {
+        // We try connect on account 10 times, now if we try 6x time, account must be locked (15 connect tentative by default on account)
+
+        // Loop login for ban
+        for (var i = 0; i < 6; i++) {
+            trapjs.addAttempt('Seby45', '112.0.0.0');
+        }
+
+        // user must be banned
+        trapjs.loginAttempt('Seby45', '112.0.0.0', function(err) {
+            assert(err != void 0);
+            assert(err.code == 'E_ACCOUNT_LOCK');
+            assert(typeof err.account.lockTime == 'number');
+            assert(err.account.lockTime > 550 && err.account.lockTime < 601);
             done();
         });
     });